Created central models file in root. Deleted separate files
